Changing IID from INT to UINT (#681)
* Changing IID from INT to UINT

* AppVeyor fix for PRs

* Fix AppVeyor script

* Maybe fix?

* Quotation marks...

* setlocal

* hmm

* debugging

* download script rewrite

* Missed a cmd line argument for DownloadFile

* yay, script clean up

* ints to uints

+++ Source/ACE.Server/Entity/Landblock.cs
@@ -889,7 +889,7 @@ namespace ACE.Server.Entity
                 return;
             throw new NotImplementedException();
             RemoveWorldObjectInternal(woGuid, false);
-            wo.ContainerId = (int)container.Guid.Full;
+            wo.ContainerId = container.Guid.Full;
 
             // We are coming off the world we need to be ready to save.
             wo.Location = null;
